Data fetched from database and displayed on ExpenseHistoryFragment.kt

+++ app/src/main/java/com/nullbyte/personalledger/model/ExpanseModel.kt
@@ -5,15 +5,15 @@ import com.google.gson.annotations.SerializedName
 
 data class ExpanseModel(
     @SerializedName("description")
-    val description: String, // Rent
+    val description: String, // description about transaction
     @SerializedName("amount")
-    val amount: Int, // 8000
+    val amount: Long, // 8000
     @SerializedName("date")
     val date: String, // 20 Jan 2020
     @SerializedName("label")
     val label: String, // Rent
     @SerializedName("peerID")
-    val peerId: String, // shubheshDwivedi@gmail.com
+    val peerId: String, // abc@xyz.com
     @SerializedName("type")
     val type: String // DEBIT
 )
